ref: move regression tests into their domain test files
Moves tests from concentrated_tick_boundary_regression.rs and
concentrated_fee_growth_regression.rs into concentrated_positions.rs,
concentrated_collect.rs, and concentrated_fees.rs respectively,
then removes the standalone regression files.
