+++ tests-integration/src/tests_reusable/mod.rs
@@ -12,11 +12,9 @@ pub mod state_sync;
 pub mod concentrated_collect;
 pub mod concentrated_create_pool;
 pub mod concentrated_failures;
-pub mod concentrated_fee_growth_regression;
 pub mod concentrated_fees;
 pub mod concentrated_positions;
 pub mod concentrated_swap;
-pub mod concentrated_tick_boundary_regression;
 pub mod concentrated_v3_fees;
 pub mod concentrated_v3_migration;
 pub mod concentrated_v3_oracle;
